requirements.txt
openwisp-controller~=0.7.post1
influxdb>=5.2,<5.3
django-notifications-hq>=1.6,<1.7
django-apptemplates>=1.4,<2.0
django-celery-email>=3.0.0,<3.1
djangorestframework>=3.11,<3.12
mac-vendor-lookup~=0.1
swapper~=1.1
django-cache-memoize~=0.1
django-nested-admin~=3.3.2
